refactor: reestrutura módulo de estatísticas e configuração via @ConfigurationProperties
- adiciona EstatisticaProperties com suporte a configuração externa (segundos)
- remove EstatisticasRepository e centraliza lógica no TransacaoRepository
- implementa cálculo de estatísticas com summaryStatistics
- ajusta controllers para injeção via construtor (sem @Autowired)
- adiciona logs com @Slf4j
- configura suporte a Java 21 preview no pom.xml
- migra application.properties para application.yml
- corrige DTO de Long para int

+++ src/main/java/itaujava10/itaujava10X/ESTATISTICAS/estatisticaDTO.java
@@ -4,13 +4,13 @@ import lombok.Getter;
 
 @Getter
 public class estatisticaDTO {
-  private final   Long count;
+  private final   int count;
     private final  Double sum;
     private final Double avg;
     private final    Double min;
     private final  Double max;
 
-    public estatisticaDTO(Long count, Double sum, Double avg, Double min, Double max) {
+    public estatisticaDTO(int count, Double sum, Double avg, Double min, Double max) {
         this.count = count;
         this.sum = sum;
         this.avg = avg;
